moved parseDate from player/details to helpers, fixed className typo in comment item
