+++ src/components/player/details.jsx
@@ -1,6 +1,7 @@
 import React            from 'react';
-import { fetchDetails } from '../../util/youtube_video_util';
 import YT_API_KEY       from '../../../config/api_key';
+import { parseDate }    from '../../helpers';
+import { fetchDetails } from '../../util/youtube_video_util';
 
 class Details extends React.Component {
 
@@ -20,31 +21,6 @@ class Details extends React.Component {
     }
   }
 
-  parseDate() {
-    const months = {
-      0: "Jan",
-      1: "Feb",
-      2: "Mar",
-      3: "Apr",
-      4: "May",
-      5: "June",
-      6: "July",
-      7: "Aug",
-      8: "Sept",
-      9: "Oct",
-      10: "Nov",
-      11: "Dec"
-    };
-    let date = new Date(this.state.details.publishedAt);
-    let newDate = "";
-
-    newDate += months[date.getMonth()] + " ";
-    newDate += date.getDate() + ", ";
-    newDate += date.getFullYear();
-
-    return newDate;
-  }
-
   addDescription() {
     if (this.state.details.description){
       let descript = this.state.details.description.slice(0,200);
@@ -56,11 +32,9 @@ class Details extends React.Component {
     return (
       <div className="details-container">
         <h5 className="details-date">
-          Published on {this.parseDate()}
+          Published on {parseDate(this.state.details.publishedAt)}
         </h5>
-        <p className="description">
-          {this.addDescription()}
-        </p>
+        {this.addDescription()}
       </div>
     );
   }
